Small refactor and few renames
